Refactor PuaraImpl methods in their own files
