+++ src/puara_config.hpp
@@ -4,6 +4,21 @@
 namespace PuaraImpl {
 enum Monitors { UART_MONITOR = 0, JTAG_MONITOR = 1, USB_MONITOR = 2 };
 
+
+  std::string get_dmi_name();
+  unsigned int get_version();
+  void set_version(unsigned int user_version);
+  std::string getIP1();
+  std::string getIP2();  
+  bool IP1_ready();
+  bool IP2_ready();
+  int unsigned getPORT1();
+  int unsigned getPORT2();
+  std::string getPORT1Str();
+  std::string getPORT2Str();
+  int unsigned getLocalPORT();
+  std::string getLocalPORTStr();
+
 inline unsigned int version = 20220906;
 inline std::string dmiName;
 inline std::string device;
