+++ src/puara_mdns.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <string_view>
+namespace PuaraImpl
+{
+  void start_mdns_service(std::string_view device_name, std::string_view instance_name);
+}
